readme: Remove VPN as a feature
The feature we provide is not VPN. You don't get actual networks that you
can connect arbitrary devices to.
As an end user, you get an ability to open terminal connections to the host
OS and to the app containers. On the backend, you get an ability to connect to
deevices from the fleet. But VPN is an implementation details, not an exposed
functionality.

Change-type: patch
Signed-off-by: Roman Mazur <roman@balena.io>